test(mesh): move TreeSync tests out of the adapter module
Move the large TreeSync adapter test suite into model_gateway/src/mesh/adapters/tests/tree_sync.rs and keep the production adapter file focused on implementation. The previous single file had grown large enough that production code and test scaffolding were hard to review together.

Keep the tests in the same crate module tree so they retain access to the private adapter state needed for repair-session assertions. This is a layout-only test refactor; behavior and coverage are intended to stay the same.
